backend/requirements/development.txt
-r base.txt

# Testing
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==4.1.0
pytest-asyncio==0.23.5
factory-boy==3.3.0
faker==24.0.0

# Code Quality
black==24.3.0
flake8==7.0.0
isort==5.13.2
pylint==3.1.0
mypy==1.9.0
django-stubs==4.2.7

# Development Tools
ipython==8.22.2
django-debug-toolbar==4.3.0
django-extensions==3.2.3